pyproject.toml
[tool.poetry]
name = "supafunc"
version = "0.2.3"
description = "Library for Supabase Functions"
authors = [
    "Joel Lee <joel@joellee.org>",
    "Andrew Smith <a.smith@silentworks.co.uk>"
]
license = "MIT"
readme = "README.md"
repository = "https://github.com/supabase-community/functions-py"

[tool.poetry.dependencies]
python = "^3.8"
httpx =  "^0.25.0"

[tool.poetry.group.dev.dependencies]
python-semantic-release = "^8.1.1"
black = "^23.9.1"
isort = "^5.12.0"
pre-commit = "^3.4.0"
pyjwt = "^2.8.0"
pytest = "^7.4.2"
pytest-cov = "^4.0.0"
unasync-cli = "^0.0.9"
pytest-asyncio = "^0.21.1"
respx = "^0.20.2"

[tool.semantic_release]
version_variables = ["supafunc/utils.py:__version__"]
version_toml = ["pyproject.toml:tool.poetry.version"]
major_on_zero = false
commit_message = "chore(release): bump version to v{version}"
build_command = "curl -sSL https://install.python-poetry.org | python - && export PATH=\"/github/home/.local/bin:$PATH\" && poetry install && poetry build"
upload_to_vcs_release = true
branch = "main"
changelog_components = "semantic_release.changelog.changelog_headers,semantic_release.changelog.compare_url"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
